Separating initial run models in coupling script
Putting the intial run of the models before the coupling loop.  This
gives us diagnostic fields in the output file that are needed for
calculating the coupling fields.  It also gives us restart files to use
so that all times through the coupling loop are identical.

Also finished implementation of the boundary layer fluxes and made sure
that the coupling fluxes are written out to the restart file (they
weren't before).