Reorganizing script and dirs to use test cases
I have modified the coupling script and the location of files and
directories to have a base directory where the script and any
documentation live, and then a subdirectory for each test case (only one
right now).  Within each test case are the files needed to create the
test case domain and the namelists for running each of the two models
for that test case.