Add script to cull MPI-SCVT triangulation
This script will take a mpas conforming mesh as input and generate text
files in the MPI-SCVT end_points.dat/triangulation.dat format as output.

It culls cells above a certain latitude based on some criteria about the
triangulation.

It can be used for two purposes for generating meshes with MPI-SCVT:
1. throwing away cells in an area you no longer want to mesh over.
2. generating a region file that only covers part of the globe.

The cell culler tool could be used for 1 and probably 2 as well, but
this tool saves a few steps.  However for very large meshes, the cell
culler may be a lot faster.